Does really belong in util.ts
